Add modules supports
